test/integration: use default GOPATH when the env var is not set
If the env var GOPATH is not set when the test runs,
now it runs `go env GOPATH` to retrieve the default value
the underlying go commands will use. I wished I could make
this command run asynchronously, but I am not familiar with
the test framework and failed to make it working with
async functions. So, like many other file operations used in
the test setup phase, I am using execFileSync.