src: restart language server automatically when its configuration changes
This change replaces the messages that tell the user to reload with automatic calls to the restart function. This way, config changes are automatically reflected - I tested it out locally and it worked pretty nicely. Most of the code to do this was already there, it was just a matter of reordering it correctly and making sure to deregister/re-register the default providers. I also added the mtime check for the language server binary as part of the config.

The only other thing that might be still missing is automatic restarts when the language server binary changes on disk, but that might be too much - probably wouldn't be intuitive for the user.

After this is merged, it will be really simple to implement #3128.

Sorry about the huge PR! It's a lot of code shuffling, and there's one function that I moved back to its original place after #3186 - sorry about that!

/cc @hyangah

Change-Id: I258dbd3a62d21da30129c08dd840d8e7ea2848e8
GitHub-Last-Rev: 29fd4d68c860e55b64eb2a41676f9dcd299f30ea
GitHub-Pull-Request: golang/vscode-go#24
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/232598
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>