src: prompt user to file an issue after gopls crashes
This change adds support for filing a gopls issue when the user restarts `gopls` or when `gopls` crashes. In a follow-up, we might be able to suggest attaching the `gopls` log.

The suggestion to file an issue after a manual restart is disabled by default for now. We can enable it once gopls is more stable. I'm moving the bulk of https://golang.org/cl/232863 here so that I can work on refactoring the env variables separately without causing a ton of merge conflicts.

Change-Id: I0c98bd526562dd50bdc7b127ddfb95f7de926075
GitHub-Last-Rev: df990d732515fa11bee78fdca45643b8cfef2cc8
GitHub-Pull-Request: golang/vscode-go#34
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/233325
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>