goLanguageServer.ts: basic implementation for gopls survey
This is the start of the gopls survey development. I put it behind a flag for now, since it's not done yet. This change sets up the basic flow of determining if users should be prompted for the survey based on their responses to prompts. A few things are still missing, namely:

- The survey prompt check should occur once every 24 hours, rather than only on activation
- The survey prompt occurs 1 hour after the user opens VS Code, but we should aim to prompt when we detect a lull in user activity

Change-Id: I861256a97265ed9ca7a717a9b07cbf0965f7d49c
GitHub-Last-Rev: 992893ec6721e0d65ef77d1fcf5595c24a558405
GitHub-Pull-Request: golang/vscode-go#22
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/231458
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>