src/goCover: clear cached coverage info on save
This is a fix for the regression in ms-vscode.Go@0.14.2

The `isDirty` bit of the TextDocument when vscode.workspace.onDidSaveTextDocument is
already false.  Adjust the condition for clearing cache not to check it.
Modification should be tracked already by trackCodeCoverageRemovalOnFileChange.

And let modifiedFiles start with an empty value rather than as undefined.

Fixes github.com/microsoft/vscode-go/issues/3252.