Use the value of GOBIN provided by `go env`.
If GOBIN is set via `go env -w GOBIN=...`, that path will
be used when vscode-go installs/updates the go tools. However,
the extension would only use the GOBIN value if set as an
environment variable prior to launching VSCode.

This commit teaches vscode-go to use the value of GOBIN
as provided by `go env`, which should be correct in all cases.