package: add missing glob dependency
https://go-review.googlesource.com/c/vscode-go/+/234020 added
new dependency on 'glob', so the package.json had to be updated.
The 'glob' module was already in the dev dependency, we couldn't
catch this during testing. Missing this dependency in the packaged
extension caused the debug adapter process to crash when loading.

We need to improve our testing to 1) add end-to-end debug testing,
2) run tests with .vsix.

Fixes golang/vscode-go#107

Change-Id: Icc7b5743def13d877069d836079c71b199b21ec9