docs: rewrite the main README
My goal in this change was to clean up the main README and make it a bit
more welcoming. The structure is now oriented towards helping people
get started and linking all of the pages in the docs/ folder.

A lot of the current content is a bit too specific and
fits better in the docs pages, so I copy-pasted a lot of text to the
relevant pages. Follow-up CLs will involve cleaning up these other pages
and smoothing out the content there.

Change-Id: I905026d1b4efb3dbad3b7c41e2f234185dee51ed
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/236057
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>