test/integration: refactor installation tests, add slow tests
Refactored the tool installation tests to follow the Mocha framework a
bit better, with setup, teardown, etc. I also added a test that should
only run in CI, when a special environment variable is set. This test
will be slow, as it download tools from the Internet. However, I think
it's worth it to have this so that we get notified earlier on about
breakages upstream.

Updates golang/vscode-go#42

Change-Id: I65edb83bf7e5b1f9374d8a9edbc8af93c6a9c660
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/236058
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>