docs/nightly: move README.md.nightly to its own directory
Hana, let me know if you don't like this change. I was just thinking
that it might be confusing for people to see 2 READMEs and 2 CHANGELOGs
at the top level. This way, we can keep this documentation separate.

Also modified the nightly README so that the information is a little
more up-to-date.

Change-Id: I6086ab0950d0fb56d3d5dfd789b45ca88985b7d7
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/236281
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>