docs: update troubleshooting documentation
Update a few broken links and create a comprehensive troubleshooting
guide. Still needs more detail, particularly since the gopls, modules,
and GOPATH documentation is still missing.

I'd love to delete the FAQ entirely -- anything worth saying
should already be in the documentation, so the information should just
be made more accessible. I tried to delete as much as possible, but some
leftovers will still have to be cleaned up in follow-up CLs. I'll
convert that section to "Known issues" which is more helpful.

Change-Id: Ie7073f9548cdbfe34af42f04d9ce5136eb1070b7
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/236757
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>