[dev.go2go] goLanguageServer.ts: add support for .go2 files with gopls
This change checks if the user has a version of go that supports go2go,
and if so, translates URIs as they go between the extension and the
language server.

It is definitely excessive to stat files up to 2 times every time they
are requested, but caching known files gets us into difficult territory
with deletions, renames, etc. I'd rather avoid it and sacrifice a little
speed.

Change-Id: I8c1e2a1973e9715c16a772bffe46fe7d3cb4b341
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/238244
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>