goLanguageServer: restart the language server 5x when it crashes
A little cleanup, and a fix for the "Never" button.
I also added handling for alternateTools for gopls, since I noticed
that it wouldn't pick up the change in setting.

I really wanted to add tests for this behavior, but constructing a
version of gopls that worked enough to initialize but then crashed
seemed to be more effort than this test is worth. Also, the timing of
waiting for the crash was just too complicated.

Change-Id: Idfb7192e875dcddc3efe55011cd1b59be57b337b
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/238657
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>