src/goCover: display coverage correctly for multiple packages
The output file from test coverage contains import paths, but when
vscode-go displays the coverage it only knows file system paths. The
code replaced by this CL used a heuristic that only got the mapping
right for a single package.

The tests do not start an editing session, but only check that the
import paths in the coverage file are properly converted to the
file system paths that vscode needs.

Change-Id: I60c8622a90134a18d0e64a239a865f0ba13ffb09
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/238697
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>