.github/workflows: remove codeql-analysis
Seems like some condition was changed after we were unforked from microsoft/vscode-go.
The test always fails with an error

"The repo on which this action is running is not opted-in to CodeQL code scanning."

Looking into an option to enable it, I found this is still in beta
and we need to sign up for waitlist.

Remove it for now. We will reconsider it when this becomes available.

Change-Id: I9388f90b87ba711d56f83b4d67bd12e394e9777a
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/239440
Reviewed-by: Rebecca Stambler <rstambler@golang.org>