src/testUtils: fix lost previous test function name
targetArgs mutates the testconfig.functions, that's an alias of lastTestConfig.
A better fix is probably to avoid the alias at the first place and store a complete
clone as the lastTestConfig, but I don't know how efficiently that can be done.

But anyway, mutating the configuration on the fly is not desirable. So, avoid it.

Fixes golang/vscode-go#269

Change-Id: I9364ac2dce9c8bb739cf76ff75b39f7ec2112f5d