package.json: update json-rpc2 and lodash for security update
This was part of https://go-review.googlesource.com/c/vscode-go/+/241041
but we are submitting it as a separate cl.

Fixes issues reported by security issues reported by npm audit.

 npm install --save json-rpc2
 npm update lodash --depth 3

json-rpc2@2.0.0 was released recently, so instead of switching to
vscode-jsonrpc, we are updating to the newest version.

Also updated the extension version 0.16.0-dev to reflect the current dev cycle.

Updates golang/vscode-go#113

Change-Id: I6a4aae5293fd505ca1ad9c39e485c3adf41c2dec
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/241557
Reviewed-by: Rebecca Stambler <rstambler@golang.org>