src/testUtils: use -json in goTest
log messages in `go test` use only the base name of the file.
When `go test` is invoked against multiple packages, we cannot tell
which package each file belongs to until the test for the package
completes and the final output including the package path is printed.
Thus, goTest buffered the test output until the final line of the package
test result becomes available and used the package path info embedded
in the last line in order to infer the full path of each log
message's file name.

This hack is no longer necessary if -json flag is used. Each entry
in the output JSON stream contains the package path.

-json flag exists since go1.9. We will no longer supports go1.8 or
older versions.

Since 1.14, go streams the test output if the `-v` flag is given.
But because of this bufferring in goTest, users couldn't benefit from
this new go1.14 feature. By eliminating this bufferring, test output
will be streamed as soon as it becomes available.

This change also fixes the file name expansion pattern because with `-v`
the log message format is different (i.e. prefixed with the test function).

Fixes golang/vscode-go#316

Note: if the log message comes from a different package (e.g. some
assertion package is used), this filename rewrite logic does not work.
That's an existing bug and without the go command's help, we cannot
fix it. A better approach would be to teach `go test` to use the full
file path in its output when requested, but it's beyond the scope of
this work.

Change-Id: I0a38bdb340e6c242ccb6eaf16a0b5af8425168fc
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/242540
Reviewed-by: Rebecca Stambler <rstambler@golang.org>