src/goModules.ts: fix test workspace in modules mode
testWorkspace may call isModSupported with the workspace URI,
which is already the root of the workspace or a module.
isModSupported unconditionally assumed the fileuril was a file
path, and worked on the parent of the workspace. That led
incorrect conclusion about modules mode, and broke tests in
various ways when modules should be enabled.

Change-Id: Ia7ba0d863e8d86f67fa27493d23ec3c686dd782e
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/242642
Reviewed-by: Rebecca Stambler <rstambler@golang.org>