src/goCover.ts: handle Windows file system and fix profile parsing
The first element of code coverage profile could be one of

 <import_path>/<go_filename> or
 <absolute_path_to_dir>/<go_filename> or
 ./<go_filename>

We considered the first case, but not the other cases.
Handle the other cases.

This CL also addresses bugs in Windows.
 - The file path can include ':' so, parsing `go list` output
   base on ':' does not work on Windows. Instead, we refactored
   getNonVendorPackages that does the same job, created a new
   getImportPathToFolder, and changed to use it.
 - While refactoring, removed the code to support go versions
   older than 1.9 - whose 'go list' results include vendored packages.
 - Fixed the cover tests to use the correct separator.

And, refactored the code in a way to avoid global variables
if possible.

Update golang/vscode-go#239.
Update #40251.

Change-Id: I60d2466c91aba95b4cb01597ad9d6ed7f9f36230
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/243177
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
Reviewed-by: Peter Weinberger <pjw@google.com>