src/goEnvironmentStatus: fix PATH mutation logic in osx
I misuderstood the comment in
https://github.com/microsoft/vscode/issues/99878#issuecomment-642808852
which resulted in that the PATH change does not apply
with the default shell args settings in OS X.

As noted in https://github.com/golang/vscode-go/issues/544#issuecomment-675897581
VS Code sets `-l` as the default shell arg.

I also attempted to run updateIntegratedTerminal after
running environmentVariableCollection.prepend. But it seems like
the onDidOpenTerminal handler is not invoked when VS Code relaunches
the terminals to apply the environmentVariableCollection changes.

Tested manually with
  "terminal.integrated.shellArgs.osx": []
and without the setting.

Fixes golang/vscode-go#544

Change-Id: Icbaaa5131893038f85e322bad1491a99f26378fd
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/251161
Reviewed-by: Suzy Mueller <suzmue@golang.org>