src/goTools.ts: recognize gofumpt, gofumports formatters
gofumpt and gofumports can be used as drop-in replacements of
gofmt and goimports, so we directed users to utilize
`"go.alternateTools"`
(https://github.com/mvdan/gofumpt#installation)

However, this is pretty convoluted and less user-friendly.
As more users are adopting gofumpt and gofumports,
promote them as officially recognized formatters.

It would be ideal if the extension allows to choose
any formatter as long as the tool follows a certain spec,
but we are not there yet. And, as we are moving towards
using the language server as the formatter, this setting
will become obsolete.

Updates golang/vscode-go#587

Change-Id: Ib1fc5c91c0d6fc7fed37132c6cbd0dfa44b6a08b
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/252922
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Suzy Mueller <suzmue@golang.org>