src/debugAdapter: send stopped events from continue on breakpoint only
The callback function for continue assumed that the reason the callback
was called was a breakpoint. However, a pause or disconnection also result
in the continue callback being called.

This change checks to see if the current thread is stopped on a
breakpoint, and will only stop the threads and notify the client if
this is true.

Fixes golang/vscode-go#149 and golang/vscode-go#172

Change-Id: Ic9c747db4cac136828d3879bbd95c2e00fe84659
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/253578
Trust: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Polina Sokolova <polina@google.com>