build: add setup_env to all.bash and update contributing.md
That can be used to install all the tools vscode-go test depends
on. We use this new command in build/Dockerfile and also,
reference in the contribution guideline.

We can also consider changing .github/workflows/* to use this
command. But I am not yet sure how to reuse it in Windows yet.

Change-Id: I328707ab4c60428e67ef43f99d5d533a6fa5f211
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/253749
Reviewed-by: Suzy Mueller <suzmue@golang.org>