src/goStatus.ts: consolidate Go status bar items
There are a number of status bar items that the extension
currently shows to the user, including the current Go version
and whether modules is enabled or not. We additionally want to
be able to let users easily check if they are using gopls, the
Go language server, since this is one of the first questions we
ask when people are having issues.

This change consolidates these items into a single item that displays
the current Go version, as well as an icon indicating if gopls is enabled.
Clicking on this item pulls up a menu to:
1. Locate the go tools and display go env
2. Choose a new go environment
3. Open the gopls trace (if applicable)
4. Open the go.mod file (if applicable)

Updates golang/vscode-go#156
Updates golang/vscode-go#525

Change-Id: I970bbe0ca64394836f62453fe5e53d57bba41f64
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/256817
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>