test: add go directive to fixtures
When running the debug adapter tests, the missing go directive
gets added to the fixtures that are used. This adds the go
directive to the go.mod files, so that when running the tests,
ther are no unexpected changes.

Change-Id: Icef235f0ca4ca9d12d2a40b3715aa875e43a6252
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/260203
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>