test: fix breakpoint tests on windows
The filepaths returned from delve (and thus the debug adapter) do not
use the native filepath separator on windows. Delve recognizes filepaths
with either separator. We can pass paths with '/' to the hitBreakpoint
function, in order to get a match.

This is a workaround to use the hitbreakpoint function as is. It may
be preferable to reimplement the function and then match the paths
ignoring the separator.

Change-Id: I8a08454ac955e5e8e3e00a657c45acc8c2bd4484
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/260729
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>