src/goMain: Add a command to copy to gopls output channel to an editor window
Intended to be used to make looking through the gopls output channel
easier than when it is just in a panel.

There doesn't seem to be a straightforward way to give it a sensible name.

It would be nice to make the Error popup go away after some period of
time.

Change-Id: I954d6bae13b7dcefd715fac23f41a0f28a0a6a28
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/263526
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>
Trust: Peter Weinberger <pjw@google.com>