src/debugAdapter: add substitutePath config for debugging
This change adds a new configuration option to both launch and
attach requests. substituePath takes an array that maps from string to string
that is used to translate paths passed to the debugger and then
back to the client.

This allows users to translate their symlinked directories to the
files that were actually used to build the binary. In addition this
can also be used for remote debugging, and when the location of the
files has moved since the program was built.

Update golang/vscode-go#622

Change-Id: I71b081d17a29655c14cd20093dc9f88867fbcc69
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/270017
Trust: Suzy Mueller <suzmue@golang.org>
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Polina Sokolova <polina@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>