test/goDebugConfiguration.test.ts: test settings.json debug config effects
The delveConfig user settings applied in settings.json, should be applied
to the debug configuration. Test that these are correctly applied.

Additionally, testFlags and buildFlags do not currently affect the debug
configuration, so we also test that to document that behavior. This behavior
is likely to change and the test should be updated when it does.

This change also reintroduces the --user-data-dir= argument to the launch
configuration for running the integration tests. This starts the test with
the default user settings, so running the tests should not be affected by
the user settings.

Update golang/vscode-go#43

Change-Id: I99c9d7f17892e6766fae731c9288880b109ba3fd
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/267217
Trust: Suzy Mueller <suzmue@golang.org>
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>