test: fix remote attach tests on windows long tests
The filepaths returned by delve will have '/' as the path separator
even on windows. One of the breakpoint tests was missing this
conversion.

The port was also not correctly set on the disconnect request.

This change does not fix all of the issues, but does keep the tests
from crashing, which should allow us to easier see the test failures.

Updates golang/vscode-go#832

Change-Id: I829ad6186001f9c665079aa69d99673068a999e8
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/276714
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
Trust: Suzy Mueller <suzmue@golang.org>