src/goLanguageServer: show language server start progress
Prevent duplicate language server restart requests
while another restart is in progress.

Allow users to cancel language server restart requests, which
results in asking users to reload the window.
We observed, when gopls is in a really bad state, it fails to respond
to the shutdown request promptly and the restart request looks hang.
In such cases, reloading the window is a quicker way to kill the
language server and recover.

Tested by using a broken language server that sleeps for a while
upon shutdown request.

Fixes golang/vscode-go#1011

Change-Id: Id1317315370ecfc8e124341a170848f7b1786335
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/276973
Reviewed-by: Suzy Mueller <suzmue@golang.org>
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>