src/welcome.ts: add welcome webview
This change is the initial change for adding a welcome page
for the extension. This adds the basic structure and a link to
open the release notes, which uses the communication channel
between the webview and the extension.

The webview is based on the microsoft webview example:
https://github.com/microsoft/vscode-extension-samples/tree/master/webview-sample

Updates golang/vscode-go#949

Change-Id: I08f63346781dcf9afb24b8d482394a0db7cf6492
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/280596
Trust: Suzy Mueller <suzmue@golang.org>
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>