docs: rewrite troubleshooting guide
The troubleshooting guide had gotten quite long, and included things
like indentation settings that aren't really troubleshooting steps. With
gopls being on by default in the near future, it's a good time to
overhaul it.

I tried to cut down and streamline the content so that it's a little
less intimidating than it used to be. I deliberately erred on the side
of removing things, because it's easier to know that's something missing
than that something's unnecessary. That said, I don't do a lot of
support for vscode-go, so I may be miscalibrated on what's important.

Hopefully it's not too gopls-centric.

A nit: I noticed along the way formatting of the Locate Configured Go
Tools command could use some cleanup; it's not clear what's what. And
it would be very nice if that command showed more of the things needed
for a bug report.

Change-Id: Id1b2ee8b7807e4c1c451fd8e477f932d69b3027f
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/284587
Trust: Heschi Kreinick <heschi@google.com>
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>