src/goMain: adjust showGoWelcomePage to show for prereleases
Previously, it showed pages only if the version string matches
the version we want to show welcome page for. And, it stored
the version in the global storage only if its major.minor.patch
verseion string is greater than the previously stored version. This
made it difficult to test the functionality with the prerelease
version.

This CL changes the logic a bit so
  - prerelease versions can exercise the code path
  - store the version info in the global state if it is different
    from the previously stored version. That means, if a user
    used v0.22.0, rolled back to v0.21.1, and reinstall v0.22.0,
    she will see the welcome page again. I hope that is a rare
    case and not too bad to show the welcome page again in this
    case.

    If a user rolls back to v0.22.0 from v0.23.0, she will not
    see the welcome page.

Change-Id: I1a0f782ab98d870dd9d90d1a9dfcc2b15af4401a
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/285837
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Rebecca Stambler <rstambler@golang.org>