src/goDebugConfiguration.ts: invoke pick process for local attach
When the processId is undefined or 0 for a local attach debug config,
we automatically invoke the quick pick to allow the user to choose
which process to attach to.

Additionally, we also throw an error when the user does not select a
process, since the debug configuration will fail.

Updates golang/vscode-go#183

Change-Id: I58c321125c8f96a3219a8ee7639779bea5dca76e
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/288953
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>