.github/workflows: use delve at master
In order to test the latest changes to delve dap, we have to
pull in the latest code. This change does this for the workflows.

go get package@version is supported in both go1.15 and in go1.16,
with go1.15 requiring GO111MODULE=on, which we already set to
download the Go tools.

Change-Id: I6d89b77edc6b028dd8e8ed9bfe4e82ccdfdaba73
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/295416
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>