test/integration/goDebug: test to change to correct goroutine when stepping
The debugger needs to switch to the correct goroutine if the threadId
in the request does not match the current goroutine. This change
checks the current state to decide if it should switch goroutines,
and then does so. This adds a test for 'next' and 'step in'
requests that switch goroutines.

Updates golang/vscode-go#118

Change-Id: I553ebbed6de1001e0e7da756278fcc717884f946
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/300769
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>