src/goTools.ts: add dlv-dap to the configured tools
dlv-dap needs to appear in the list of missing tools in order for
the user get the prompt to install.

Additionally, if they already have dlv-dap installed, we do want
dlv-dap to appear as a tool to update, since we are building from
master.

If the user has declined to install, do not attempt to prompt,
because it will do nothing. Instead, continue with the debugging
session without attempting to install so the error will show up
later.

Updates golang/vscode-go#794

Change-Id: If5922c950fde621d4c666e06bcc149413674ed9d
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/303235
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>